pyproject.toml
[tool.poetry]
name = "my-site"
version = "0.1.0"
description = ""
authors = ["Azamat Baltaev <1210044@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.6.0"
attrs = "23.1.0"
build = "0.10.0"
CacheControl = "0.12.14"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "3.0.1"
cleo = "2.0.1"
crashtest = "0.4.1"
cryptography = "41.0.1"
distlib = "0.3.6"
Django = "4.1.5"
django-debug-toolbar = "4.1.0"
django-filter = "23.2"
djangorestframework = "3.14.0"
docutils = "0.20.1"
drf-spectacular = "0.26.2"
dulwich = "0.21.5"
filelock = "3.12.2"
flake8 = "6.0.0"
flake8-docstrings = "1.7.0"
gunicorn = "20.1.0"
html5lib = "1.1"
idna = "3.4"
importlib-metadata = "6.7.0"
inflection = "0.5.1"
installer = "0.7.0"
jeepney = "0.8.0"
jsonschema = "4.17.3"
keyring = "23.13.1"
lockfile = "0.12.2"
mccabe = "0.7.0"
more-itertools = "9.1.0"
msgpack = "1.0.5"
packaging = "23.1"
pexpect = "4.8.0"
Pillow = "9.4.0"
pkginfo = "1.9.6"
platformdirs = "3.8.0"
poetry = "1.5.1"
poetry-core = "1.6.1"
poetry-plugin-export = "1.4.0"
ptyprocess = "0.7.0"
pycodestyle = "2.10.0"
pycparser = "2.21"
pydocstyle = "6.3.0"
pyflakes = "3.0.1"
pyproject_hooks = "1.0.0"
pyrsistent = "0.19.3"
pytz = "2023.3"
PyYAML = "6.0"
rapidfuzz = "2.15.1"
requests = "2.28.2"
requests-toolbelt = "1.0.0"
SecretStorage = "3.3.3"
sentry-sdk = "1.25.1"
shellingham = "1.5.0.post1"
six = "1.16.0"
snowballstemmer = "2.2.0"
sqlparse = "0.4.3"
tomli = "2.0.1"
tomlkit = "0.11.8"
trove-classifiers = "2023.5.24"
uritemplate = "4.1.1"
urllib3 = "1.26.14"
virtualenv = "20.23.1"
webencodings = "0.5.1"
zipp = "3.15.0"
jaraco-classes = "3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
